+++ ui/components/modals/base/example.jsx
@@ -55,6 +55,17 @@ export let Modal = (props) => (
     className={classNames('slds-modal slds-fade-in-open', props.className)}
   >
     <div className="slds-modal__container">
+      {
+        props.closeButton !== false ? (
+          <ButtonIcon
+            className="slds-modal__close slds-button_icon-inverse"
+            iconClassName="slds-button__icon_large"
+            symbol="close"
+            assistiveText="Cancel and close"
+            disabled={props.closeButtonDisabled}
+          />
+        ) : null
+      }
       {props.children}</div>
   </section>
 );
@@ -71,40 +82,11 @@ Modal.defaultProps = {
   isFullSize: false,
 };
 
-export let ModalHeader = (props) => {
-  const closeIcon = (
-    <ButtonIcon
-      className="slds-modal__close"
-      iconClassName="slds-button__icon_large"
-      symbol="close"
-      assistiveText="Cancel and close"
-      disabled={props.closeButtonDisabled}
-    />
-  );
-  let header = [];
-  if(props.headless) {
-    header.push(closeIcon);
-  }
-  else {
-    React.Children.forEach(props.children, (child, index) => {
-      header.push(child);
-      if(index === 0) {
-        (props.closeButton !== false) && header.push(closeIcon);
-      }
-    });
-  }
-  return (<div className={classNames('slds-modal__header slds-text-align_left', props.className)}>
-    {header}
-  </div>);
-  };
-
-ModalHeader.propTypes = {
-  headless: PropTypes.bool
-};
-
-ModalHeader.defaultProps = {
-  headless: false
-};
+export let ModalHeader = (props) => (
+  <div className={classNames('slds-modal__header', props.className)}>
+    {props.children}
+  </div>
+);
 
 export let ModalHeaderDeprecated = (props) => (
   <div className={classNames('slds-modal__header', props.className)}>
@@ -279,7 +261,6 @@ export let FullHeadless = (props) => (
       aria-labelledby="modal-heading-01"
       isFullSize={props.size === 'full'}
     >
-      <ModalHeader headless className="slds-modal__header_empty" />
       <ModalContent className={classNames(
         props.size === 'full' && 'slds-modal_full-content slds-p-around_medium slds-modal__content_headless'
       )}
@@ -375,7 +356,6 @@ export let Directional = (props) => (
 export let Headless = (props) => (
   <Backdrop>
     <Modal aria-label="Meaningful description of the modal content">
-      <ModalHeader headless className="slds-modal__header_empty" />
       <ModalContent className="slds-p-around_medium slds-modal__content_headless">
         <p>
           Sit nulla est ex deserunt exercitation anim occaecat. Nostrud ullamco
@@ -465,7 +445,6 @@ export let HiddenFooter = (props) => (
 export let HeadlessAndFootless = (props) => (
   <Backdrop>
     <Modal aria-labelledby="modal-content-id-1">
-      <ModalHeader headless className="slds-modal__header_empty" />
       <ModalContent className="slds-modal__content_headless slds-modal__content_footless slds-p-around_medium">
         <p>
           Sit nulla est ex deserunt exercitation anim occaecat. Nostrud ullamco
@@ -491,7 +470,7 @@ export let HeadlessAndFootless = (props) => (
 export let HeadlessAndFootlessDeprecated = (props) => (
   <Backdrop>
     <Modal aria-labelledby="modal-content-id-1">
-      <ModalHeader headless className="slds-modal__header_empty" />
+      <ModalHeader className="slds-modal__header_empty" />
       <ModalContent className="slds-p-around_medium">
         <p>
           Sit nulla est ex deserunt exercitation anim occaecat. Nostrud ullamco
@@ -526,7 +505,7 @@ export const Menu = () => {
         <ModalHeader>
           <Heading
             id={uniqueIDHeading}
-            className="slds-modal__title slds-hyphenate"
+            className="slds-modal__title slds-hyphenate" tabindex="-1"
           >
             Modal header
           </Heading>
@@ -557,8 +536,8 @@ export const Menu = () => {
 
 export let DisabledClose = (props) => (
   <Backdrop>
-    <Modal aria-labelledby="modal-heading-01">
-      <ModalHeader closeButtonDisabled>
+    <Modal aria-labelledby="modal-heading-01" closeButtonDisabled>
+      <ModalHeader>
         <h1 id="modal-heading-01" className="slds-modal__title slds-hyphenate" tabindex="-1">
           Modal header
         </h1>